// src/classes/Table/types.d.ts
export interface TableConstructor<Name extends string, PartitionKey extends A.Key, SortKey extends A.Key | null> {
    name: Name;
    alias?: string | null;
    partitionKey: PartitionKey;
    sortKey?: SortKey;
    entityField?: boolean | string;
    attributes?: TableAttributes;
    indexes?: TableIndexes;
    autoExecute?: boolean;
    autoParse?: boolean;
    DocumentClient?: DynamoDBDocumentClient;
    entities?: {};
    removeNullAttributes?: boolean;
}
export declare type DynamoDBTypes = 'string' | 'boolean' | 'number' | 'bigint' | 'list' | 'map' | 'binary' | 'set';
export declare type DynamoDBKeyTypes = 'string' | 'number' | 'bigint' | 'binary';
export interface executeParse {
    execute?: boolean;
    parse?: boolean;
}
export interface TableAttributeConfig {
    type: DynamoDBTypes;
    setType?: DynamoDBKeyTypes;
}
export interface TableAttributes {
    [attr: string]: DynamoDBTypes | TableAttributeConfig;
}
export interface ParsedTableAttribute {
    [attr: string]: TableAttributeConfig & {
        mappings: {};
    };
}
export interface TableIndexes {
    [index: string]: {
        partitionKey?: string;
        sortKey?: string;
    };
}